tech(TTB-060): Moved and renamed Board to GameBoard
