+++ client/src/components/GameBoard/index.js
@@ -1,0 +1,3 @@
+import GameBoard from "./GameBoard";
+
+export default GameBoard;
